+++ src/module/helpers/rolldialog-sf.ts
@@ -1,5 +1,6 @@
+import { maxBy, minBy } from 'lodash'
 import { IronswornActor } from '../actor/actor'
-import { createStarforgedMoveRollChat } from '../chat/chatrollhelpers'
+import { createStarforgedMoveRollChat, sfNextOracles } from '../chat/chatrollhelpers'
 import { IronswornItem } from '../item/item'
 import { SFMoveDataProperties } from '../item/itemtypes'
 import { IronswornSettings } from './settings'
@@ -76,14 +77,23 @@ function callback(opts: { actor: IronswornActor; move: IronswornItem; mode: stri
 
 async function rollAndCreateChatMessage(opts: { actor: IronswornActor; move: IronswornItem; mode: string; stats: string[]; bonus: number }) {
   const { actor, move, mode, stats, bonus } = opts
-  if (mode !== 'Stat') {
+
+  const normalizedStats = stats.map((x) => x.toLowerCase())
+  let usedStat = normalizedStats[0]
+  if (mode === 'Best of' || mode === 'Worst of') {
+    const statMap = {}
+    for (const x of normalizedStats) {
+      statMap[x] = actor.data.data[x]
+    }
+    const fn = mode === 'Best of' ? maxBy : minBy
+    usedStat = fn(Object.keys(statMap), (x) => statMap[x]) ?? stats[0]
+  } else if (mode !== 'Stat') {
     console.log({ actor, move, mode, stats, bonus })
-    return // TODO: implement best/worst/all of
+    return // TODO: all of
   }
 
   let actionExpr = 'd6'
-  const stat = stats[0]
-  if (stat) actionExpr += ` + @${stat.toLowerCase()}`
+  if (usedStat) actionExpr += ` + @${usedStat}`
   if (bonus) actionExpr += ` + ${bonus}`
   const data = {
     ...actor?.getRollData(),
@@ -95,13 +105,18 @@ async function rollAndCreateChatMessage(opts: { actor: IronswornActor; move: Iro
     actor,
     move,
     mode,
-    stats,
+    stats: normalizedStats,
+    usedStat,
     bonus,
   })
 }
 
 async function createDataforgedMoveChat(move: IronswornItem) {
-  const content = await renderTemplate('systems/foundry-ironsworn/templates/chat/sf-move.hbs', { move })
+  const params = {
+    move,
+    nextOracles: await sfNextOracles(move),
+  }
+  const content = await renderTemplate('systems/foundry-ironsworn/templates/chat/sf-move.hbs', params)
   ChatMessage.create({
     speaker: ChatMessage.getSpeaker(),
     content,