+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## In progress
 
+- Provide oracle-rolling buttons on move chat cards ([#287](https://github.com/ben/foundry-ironsworn/pull/287))
+- Implement "best of" and "worst of" move rolling (also [#287](https://github.com/ben/foundry-ironsworn/pull/287))
+
 ## 1.10.40
 
 - Fix location token and oracle bugs ([#284](https://github.com/ben/foundry-ironsworn/pull/284))